+++ MasterOk/Controllers/UsersController.cs
@@ -1,5 +1,4 @@
-﻿#nullable disable
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
@@ -9,11 +8,9 @@ using Microsoft.EntityFrameworkCore;
 using MasterOk.Data;
 using MasterOk.Models.ModelDataBase;
 using MasterOk.Models.FilterSortViewModels;
-using Microsoft.AspNetCore.Authorization;
 
 namespace MasterOk.Controllers
 {
-    [Authorize(Roles = "user")]
     public class UsersController : Controller
     {
         private readonly DataBaseContext _context;
@@ -26,7 +23,7 @@ namespace MasterOk.Controllers
         // GET: Users
         public async Task<IActionResult> Index(ESortModelUsers sort)
         {
-            IQueryable<User> dataBaseContext = _context.Users;
+            IQueryable<User> dataBaseContext = _context.Users.Include(u => u.Role);
 
             dataBaseContext = sort switch
             {
@@ -44,8 +41,11 @@ namespace MasterOk.Controllers
                 ESortModelUsers.NameDesc => dataBaseContext.OrderByDescending(s => s.FirstLastNameStaff),
                 ESortModelUsers.NumberPhoneAsc => dataBaseContext.OrderBy(s => s.NumberPhoneStaff),
                 ESortModelUsers.NumberPhoneDesc => dataBaseContext.OrderByDescending(s => s.NumberPhoneStaff),
+                ESortModelUsers.RoleAsc => dataBaseContext.OrderBy(s => s.Role.TitleRole),
+                ESortModelUsers.RoleDesc => dataBaseContext.OrderByDescending(s => s.Role.TitleRole),
                 _ => dataBaseContext
             };
+
 
             return View(new SortViewModelUsers
             {
@@ -57,12 +57,14 @@ namespace MasterOk.Controllers
         // GET: Users/Details/5
         public async Task<IActionResult> Details(int? id)
         {
-            if (id == null)
+            if (id == null || _context.Users == null)
             {
                 return NotFound();
             }
 
-            var user = await _context.Users.FirstOrDefaultAsync(m => m.Id == id);
+            var user = await _context.Users
+                .Include(u => u.Role)
+                .FirstOrDefaultAsync(m => m.Id == id);
             if (user == null)
             {
                 return NotFound();
@@ -74,6 +76,7 @@ namespace MasterOk.Controllers
         // GET: Users/Create
         public IActionResult Create()
         {
+            ViewData["RoleId"] = new SelectList(_context.Roles, "Id", "TitleRole");
             return View();
         }
 
@@ -82,7 +85,7 @@ namespace MasterOk.Controllers
         // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("Id,EmailUser,LoginUser,PasswordUser,FirstLastNameStaff,Age,NumberPhoneStaff, ActiveUser")] User user)
+        public async Task<IActionResult> Create([Bind("Id,EmailUser,LoginUser,PasswordUser,FirstLastNameStaff,Age,NumberPhoneStaff,ActiveUser,RoleId")] User user)
         {
             if (ModelState.IsValid)
             {
@@ -90,13 +93,14 @@ namespace MasterOk.Controllers
                 await _context.SaveChangesAsync();
                 return RedirectToAction(nameof(Index));
             }
+            ViewData["RoleId"] = new SelectList(_context.Roles, "Id", "TitleRole", user.RoleId);
             return View(user);
         }
 
         // GET: Users/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
-            if (id == null)
+            if (id == null || _context.Users == null)
             {
                 return NotFound();
             }
@@ -106,6 +110,7 @@ namespace MasterOk.Controllers
             {
                 return NotFound();
             }
+            ViewData["RoleId"] = new SelectList(_context.Roles, "Id", "TitleRole", user.RoleId);
             return View(user);
         }
 
@@ -114,7 +119,7 @@ namespace MasterOk.Controllers
         // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("Id,EmailUser,LoginUser,PasswordUser,FirstLastNameStaff,Age,NumberPhoneStaff,ActiveUser")] User user)
+        public async Task<IActionResult> Edit(int id, [Bind("Id,EmailUser,LoginUser,PasswordUser,FirstLastNameStaff,Age,NumberPhoneStaff,ActiveUser,RoleId")] User user)
         {
             if (id != user.Id)
             {
@@ -141,18 +146,21 @@ namespace MasterOk.Controllers
                 }
                 return RedirectToAction(nameof(Index));
             }
+            ViewData["RoleId"] = new SelectList(_context.Roles, "Id", "TitleRole", user.RoleId);
             return View(user);
         }
 
         // GET: Users/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
-            if (id == null)
+            if (id == null || _context.Users == null)
             {
                 return NotFound();
             }
 
-            var user = await _context.Users.FirstOrDefaultAsync(m => m.Id == id);
+            var user = await _context.Users
+                .Include(u => u.Role)
+                .FirstOrDefaultAsync(m => m.Id == id);
             if (user == null)
             {
                 return NotFound();
@@ -166,15 +174,23 @@ namespace MasterOk.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
+            if (_context.Users == null)
+            {
+                return Problem("Entity set 'DataBaseContext.Users'  is null.");
+            }
             var user = await _context.Users.FindAsync(id);
-            _context.Users.Remove(user);
+            if (user != null)
+            {
+                _context.Users.Remove(user);
+            }
+
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
         }
 
         private bool UserExists(int id)
         {
-            return _context.Users.Any(e => e.Id == id);
+            return (_context.Users?.Any(e => e.Id == id)).GetValueOrDefault();
         }
     }
 }
